move pagination component out of topics folder
