+++ src/app/components/Pagination/index.test.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render } from '../../../components/react-testing-library-with-providers';
+import { render } from '../react-testing-library-with-providers';
 import PaginationComponent from '.';
 
 // eslint-disable-next-line react/prop-types
